+++ CHANGELOG.md
@@ -131,6 +131,7 @@
 - Bug #132: Update related tests classes to ensure immutability and add missing tests (@terabytesoftw)
 - Enh #133: Add `HasPatternContentUnits` trait and corresponding tests for managing SVG `patternContentUnits` attribute (@terabytesoftw)
 - Enh #134: Add `HasPatternTransform` trait and corresponding tests for managing SVG `patternTransform` attribute (@terabytesoftw)
+- Enh #135: Add `HasPatternUnits` trait and corresponding tests for managing SVG `patternUnits` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024